+++ docs/CHANGELOG.rst
@@ -6,6 +6,11 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+
 ==================
 9.0.0 - 2018-05-15
 ==================